Fix Clippy warning and formatting
- Fixed `clippy::unnecessary_unwrap` in `src/cli.rs` by using `assert!(result.is_ok(), ...)` instead of manual check.
- Ran `cargo fmt` to fix indentation issues.
- Restored `src/cli.rs` and `src/repl.rs` which were briefly missing in the context but needed for compilation.
- Verified all tests pass.
