+++ src/lib.rs
@@ -35,27 +35,23 @@
 //! # Module Guide
 //!
 //! * [`ast`]: Abstract Syntax Tree definitions.
-//! * [`cli`]: CLI command implementations.
 //! * [`codegen`]: Rust code generation logic.
 //! * [`errors`]: Greek-native error messages and diagnostics.
 //! * [`grammar`]: PEG parser.
 //! * [`highlight`]: Semantic syntax highlighting.
 //! * [`morphology`]: Word analysis, lexicon, and participle parsing.
 //! * [`parser`]: AST Builder and parsing logic.
-//! * [`repl`]: Interactive REPL.
 //! * [`report`]: Report generation and statistics.
 //! * [`semantic`]: The slot-based assembler and semantic analysis.
 //! * [`text`]: Text utilities and normalization.
 
 pub mod ast;
-pub mod cli;
 pub mod codegen;
 pub mod errors;
 pub mod grammar;
 pub mod highlight;
 pub mod morphology;
 pub mod parser;
-pub mod repl;
 pub mod report;
 pub mod semantic;
 pub mod text;
